feature: disable profile screen when anonymous
Also refactor of SignIn screen
