+++ app/src/main/java/com/github/se/cyrcle/ui/theme/molecules/TopAppBar.kt
@@ -1,6 +1,5 @@
 package com.github.se.cyrcle.ui.theme.molecules
 
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material3.CenterAlignedTopAppBar
@@ -14,7 +13,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.unit.dp
 import com.github.se.cyrcle.ui.navigation.NavigationActions
 import com.github.se.cyrcle.ui.theme.Typography
 
@@ -43,7 +41,7 @@ fun TopAppBar(navigationActions: NavigationActions, title: String, testTag: Stri
               containerColor = MaterialTheme.colorScheme.primaryContainer,
               actionIconContentColor = MaterialTheme.colorScheme.primaryContainer,
               scrolledContainerColor = MaterialTheme.colorScheme.primaryContainer),
-      modifier = Modifier.padding(vertical = 8.dp).testTag(testTag),
+      modifier = Modifier.testTag(testTag),
       navigationIcon = {
         IconButton(
             onClick = { navigationActions.goBack() },
